Add RBAC tests
This proof of concept leverages the nine default personas available
through the pre-provisioned credentials model[1] to demonstrate a potential
framework for testing default policies. In this example, an abstract
base class is created that helps set up credentials and outlines every
policy that needs to be tested, then nine subclasses are created to test
every persona. Each test represents one policy rule, and some tests make
multiple requests in order to test the policy from different approaches,
for example, to check what happens if a different domain is specified,
or what happens if the resource does not exist.

[1] https://github.com/cmurphy/tempest/commit/63a176b1bf79fd6b8cebe66d2b0e153634427488

Change-Id: Icb5317b9297230490bd783fe9b07c8db244c06f8